+++ CHANGELOG.md
@@ -15,7 +15,7 @@
   - Added variables for width sub-modifiers to form component
   - Added size modifier support for textarea and multiple selects
   - Added renderer option for seach component
-  - Added scrollspynav to scrollspy component
+  - Added scrollspy-nav to scrollspy component
   - Auto-initialize dynamic components using MutationObserver if supported by browser
   - Changed box-sizing for scrollable-box
   - Fixed link-muted